requirements.txt
ccxt==3.0.77
uvicorn==0.22.0
fastapi==0.95.2
redis>=5.0.12
pandas==2.2.3
numpy>=1.27.4
python-dotenv==1.0.1
scikit-learn>=1.3.2
joblib>=1.3.2
streamlit==1.36.0
plotly==5.22.0
requests==2.32.3
ta==0.10.2         # Indicadores técnicos (RSI, MACD, EMA…)
lightgbm==4.1.0    # Opcional, mejor que RF para stacking

# IA / Deep Learning
torch>=2.3.0
torchvision>=0.18.0

# RL
stable-baselines3>=2.3.2
tensorboard>=2.17.0
gymnasium>=0.28.1,<1.0
shimmy>=1.3.0

# NLP / Sentiment
transformers>=4.41.0
# opcional para mejores despliegues de transformers (CPU/GPU)
accelerate>=0.33.0

# RSS + fechas para ingesta de noticias
feedparser>=6.0.10
python-dateutil>=2.8.2
